fix: Resolve all compilation errors across all apps and core package
Core Package Fixes:
- Updated connectivity_plus to ^6.0.0 (new List<ConnectivityResult> API)
- Added missing AppTypography.heading3 getter
- Added missing AppSpacing.radiusSm and radiusMd getters
- Fixed AppLocalizations._strings return type (Object -> dynamic)
- Removed unused fields in notification_handler.dart and push_notification_service.dart
- Fixed deep_link_service.dart to use getInitialAppLinkString()
- Removed unused isDark variables in 13 UI components
- Fixed dead null-aware expression in app_search_bar.dart

Customer App:
- Added flutter_stripe, firebase_auth, cloud_functions dependencies
- Fixed membership_screen.dart to use Stripe.instance directly
- Added flutter_stripe import

Driver App:
- Added firebase_auth, firebase_storage, image_picker dependencies

Partner App:
- Added firebase_auth, firebase_storage dependencies

All apps now compile with no errors!

+++ tayyebgo_monorepo/packages/tayyebgo_core/lib/src/services/delivery_fee_service.dart
@@ -54,7 +54,7 @@ class DeliveryFeeService {
     required int timeOfDay,
     required double demand,
   }) {
-    double fee = baseFee;
+    double fee = baseFee.toDouble();
     fee += distance * perKmRate;
 
     final zoneMultiplier = zoneRates[zone] ?? 1.0;
